+++ apps/dashboard/src/components/Providers.tsx
@@ -1,8 +1,10 @@
 "use client";
-import type { ReactNode } from "react";
+import { useEffect, type ReactNode } from "react";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { Theme } from "@radix-ui/themes";
 import { ThemeProvider } from "next-themes";
+import { useImagesStore } from "../stores/imagesStore";
+import { useZoomStore } from "../stores/zoomStore";
 
 const queryClient = new QueryClient();
 
@@ -11,6 +13,11 @@ interface ProvidersProps {
 }
 
 export function Providers({ children }: ProvidersProps) {
+  useEffect(() => {
+    void useImagesStore.persist.rehydrate();
+    void useZoomStore.persist.rehydrate();
+  }, []);
+
   return (
     <QueryClientProvider client={queryClient}>
       <ThemeProvider attribute="class">
